Fix billable would be overwritten to false even if no change occured
The radio buttons would emit a string but the hydration of the form would set billable to a bool.
When the form was sent a check to equality against 'true'  was performed but when the user did not
change the value the hydrated bool value would still be set to the equality would be false.